finish gametime conversion

+++ Source/Player/PlayerInventory.cs
@@ -113,7 +113,7 @@ internal class PlayerInventory(ILevelManager levelManager) {
       float barHeight = 8f;
       Vector2 barPos = Position + new Vector2(-barWidth / 2f, -80f);
 
-      float progress = 1f - (gun.ReloadTimer / gun.PublicStats.ReloadTime);
+      float progress = 1.0f - ((float) (gun.ReloadTimer / gun.PublicStats.ReloadTime));
 
       spriteBatch.Draw(whitePixel, new Rectangle((int) barPos.X, (int) barPos.Y, (int) barWidth, (int) barHeight), Color.DarkGray * 0.8f);
       spriteBatch.Draw(whitePixel, new Rectangle((int) barPos.X, (int) barPos.Y, (int) (barWidth * progress), (int) barHeight), Color.White);
